Switch to a custom Uri sexp and use compare
We need Uri.compare for ppx_inline_test, and the sexp format for Uri
is annoyingly verbose, so use a custom simpler sexp format and
continue using Uri.compare.

Currently this stuff is all in dsn.ml since we don't use it anywhere
else but it should be moved if we ever do.